# openapi.json
{"openapi":"3.0.1","swagger": "2.0","info":{"version": "v1","title":"Cleaning Application","description":"TEAM 1 PROJECT"},"servers":[{"url":"http://localhost:8080"}],"tags":[{"name":"Address API","description":"Endpoint for operations with addresses"},{"name":"CommercialProposal API","description":"Endpoint for operations with commercial proposals"},{"name":"User API","description":"Endpoint for operations with users (customers/staff)"},{"name":"Comment API","description":"Endpoint for operations with comments of orders"},{"name":"Order API","description":"Endpoint for operations with orders"}],"paths":{"/users":{"put":{"tags":["User API"],"summary":"Change user","description":"Change user","operationId":"updateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"post":{"tags":["User API"],"summary":"Add user","description":"Add user","operationId":"addUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/orders":{"put":{"tags":["Order API"],"summary":"Change order if status NOT_PROCESSED","description":"Change order if status NOT_PROCESSED","operationId":"editOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["Order API"],"summary":"Create new order","description":"Create new order","operationId":"createOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/orders/status":{"put":{"tags":["Order API"],"summary":"Change order status","description":"Change order status","operationId":"changeStatus","parameters":[{"name":"status","in":"query","required":true,"schema":{"type":"string","enum":["NOT_VERIFIED","VERIFIED","NOT_STARTED","IN_PROGRESS","DONE","CANCELLED"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/commercial-proposals":{"put":{"tags":["CommercialProposal API"],"summary":"Change commercial proposal","description":"Change commercial proposal","operationId":"editProposal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommercialProposalDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommercialProposalDto"}}}}}},"post":{"tags":["CommercialProposal API"],"summary":"Create new commercial proposal","description":"Create new commercial proposal","operationId":"addProposal","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommercialProposalDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommercialProposalDto"}}}}}}},"/comments":{"put":{"tags":["Comment API"],"summary":"Change comment","description":"Change comment","operationId":"editProposal_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDto"}}},"required":true},"responses":{"200":{"description":"OK"}}},"post":{"tags":["Comment API"],"summary":"Create new comment for order","description":"Create new comment for order","operationId":"createComment","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommentDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/addresses":{"get":{"tags":["Address API"],"summary":"Get all user addresses","description":"Get all user addresses","operationId":"getUserAddresses","parameters":[{"name":"userDto","in":"query","required":true,"schema":{"$ref":"#/components/schemas/UserDto"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AddressDto"}}}}}}},"put":{"tags":["Address API"],"summary":"Change address","description":"Change address","operationId":"editAddress","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/addresses/{userId}":{"post":{"tags":["Address API"],"summary":"Create new address for user","description":"Create new address for user","operationId":"createAddress","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/users/{id}":{"get":{"tags":["User API"],"summary":"Get user by id","description":"Get user by id","operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}},"delete":{"tags":["User API"],"summary":"Delete user","description":"Delete user","operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/users/by-email":{"get":{"tags":["User API"],"summary":"Get user by email","description":"Get user by email","operationId":"getUserByEmail","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}}}},"/orders/{id}":{"get":{"tags":["Order API"],"summary":"Get order by id","description":"Get order by id","operationId":"getOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/OrderDto"}}}}}}},"/commercial-proposals/{id}":{"get":{"tags":["CommercialProposal API"],"summary":"Get commercial proposal by id","description":"Get commercial proposal by id","operationId":"getProposalById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommercialProposalDto"}}}}}},"delete":{"tags":["CommercialProposal API"],"summary":"Delete commercial proposal","description":"Delete commercial proposal","operationId":"deleteProposal","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/comments/{id}":{"get":{"tags":["Comment API"],"summary":"Get comment by id","description":"Get comment by id","operationId":"getComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CommentDto"}}}}}},"delete":{"tags":["Comment API"],"summary":"Delete comment","description":"Delete comment","operationId":"deleteComment","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/addresses/{id}":{"get":{"tags":["Address API"],"summary":"Get address by id","description":"Get address by id","operationId":"getAddress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AddressDto"}}}}}},"delete":{"tags":["Address API"],"summary":"Delete address","description":"Delete address","operationId":"deleteAddress","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserDto":{"required":["email","name","password","role"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"maxLength":2147483647,"minLength":12,"pattern":"^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d).+$","type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["User","Employee","Admin"]}}},"Address":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"city":{"type":"string"},"street":{"type":"string"},"houseNumber":{"type":"string"},"flatNumber":{"type":"integer","format":"int32"}}},"Comment":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"text":{"type":"string"},"date":{"type":"string","format":"date-time"},"rate":{"type":"integer","format":"int32"}}},"CommercialProposal":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"}}},"OrderDto":{"required":["address","client","comment","commercialProposals","executors","orderStatus"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"price":{"minimum":0,"type":"number","format":"double"},"orderTime":{"type":"string","format":"date-time"},"creationTime":{"type":"string","format":"date-time"},"client":{"$ref":"#/components/schemas/User"},"executors":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/User"}},"comment":{"$ref":"#/components/schemas/Comment"},"address":{"$ref":"#/components/schemas/Address"},"orderStatus":{"type":"string","enum":["NOT_VERIFIED","VERIFIED","NOT_STARTED","IN_PROGRESS","DONE","CANCELLED"]},"commercialProposals":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/CommercialProposal"}}}},"User":{"required":["email","name","password","role"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"password":{"maxLength":2147483647,"minLength":12,"pattern":"^(?=.*[A-Z])(?=.*[a-z])(?=.*\\d).+$","type":"string"},"email":{"type":"string"},"role":{"type":"string","enum":["User","Employee","Admin"]}}},"CommercialProposalDto":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"price":{"type":"number","format":"double"}}},"CommentDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"text":{"type":"string"},"date":{"type":"string","format":"date-time"},"rate":{"type":"integer","format":"int32"}}},"AddressDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"city":{"type":"string"},"street":{"type":"string"},"houseNumber":{"type":"string"},"flatNumber":{"type":"integer","format":"int32"}}}}}}